requirements.txt
Flask>=2.0
PyYAML>=6.0
Markdown>=3.0
gunicorn>=20.0
